# swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cocktail Ordering System API",
    "version": "1.0.0",
    "description": "A complete backend API for a cocktail ordering system with Express.js, Node.js, and MongoDB. Features JWT authentication, Paystack payment integration, and comprehensive admin management.",
    "contact": {
      "name": "Clyp AI",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
...
  },
...
  "paths": {
...
    "/email/send-otp": {
      "post": {
        "summary": "Send OTP for email validation (Public endpoint)",
        "tags": [
          "Email Validation"
        ],
        "description": "Send a One-Time Password (OTP) to the specified email address for validation.\nThe OTP will be sent via Gmail SMTP and expires in 10 minutes.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "email_verification",
                      "password_reset",
                      "admin_registration"
                    ],
                    "default": "email_verification",
                    "description": "Purpose of the OTP"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "string",
                      "description": "OTP expiration time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or email already verified"
          },
          "500": {
            "description": "Internal server error or email sending failed"
          }
        }
      }
    },
    "/email/verify-otp": {
      "post": {
        "summary": "Verify OTP for email validation (Public endpoint)",
        "tags": [
          "Email Validation"
        ],
        "description": "Verify the OTP token sent to the email address.\nIf valid, the email will be marked as verified.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "otp"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "otp": {
                    "type": "string",
                    "example": "123456",
                    "description": "OTP token received via email"
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "email_verification",
                      "password_reset",
                      "admin_registration"
                    ],
                    "default": "email_verification",
                    "description": "Purpose of the OTP"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "verifiedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid OTP or validation error"
          },
          "404": {
            "description": "OTP not found or expired"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/email/resend-otp": {
      "post": {
        "summary": "Resend OTP for email validation (Public endpoint)",
        "tags": [
          "Email Validation"
        ],
        "description": "Resend OTP to the specified email address.\nThis will invalidate any existing OTP for the email.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "purpose": {
                    "type": "string",
                    "enum": [
                      "email_verification",
                      "password_reset",
                      "admin_registration"
                    ],
                    "default": "email_verification"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP resent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}